# api/public/docs/openapi.yml
openapi: 3.1.0
info:
  title: 'Constellation APIs'
  license:
    name: 'GNU AFFERO GENERAL PUBLIC LICENSE'
    url: 'https://www.gnu.org/licenses/agpl-3.0.en.html'
  version: '2.0'
servers:
  -
    url: 'http://localhost:8080/api'
    description: 'The local environment.'
paths:
...
  /v1/service_instance_dependencies:
    get:
      tags:
        - 'Service instance dependencies'
      summary: 'Get a listing of the service instance dependencies'
      description: 'Get all service instance dependenciess.'
      operationId: getServiceInstanceDependenciess
      parameters:
        -
          $ref: '#/components/parameters/base-filter-per-page'
        -
          $ref: '#/components/parameters/base-filter-page'
        -
          $ref: '#/components/parameters/base-filter-sort'
        -
          $ref: '#/components/parameters/base-filter-q'
        -
          name: 'filter[id]'
          in: query
          description: 'Filter by service instance dependency id.'
          schema:
            type: integer
      responses:
        '200':
          description: 'Service instance dependenciess list'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/resource-service-instance-dependencies' } }
                  message: { type: string }
                  total: { type: integer }
                type: object
    post:
      tags:
        - 'Service instance dependencies'
      summary: 'Store an service instance dependency'
      description: 'Store an service instance dependency.'
      operationId: storeServiceInstanceDependency
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request-create-service-instance-dependency'
      responses:
        '200':
          description: 'Created serviceinstancedependencies data.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/resource-service-instance-dependencies' }
                  message: { type: string }
                type: object
        '422':
          description: 'Invalid inputs.'
          content:
            application/json:
              schema:
                properties:
                  errors: { properties: { '{name-of-invalid-field}': { type: array, items: { description: 'Error message', type: string } } }, type: object }
                  message: { type: string }
                type: object
  '/v1/service_instance_dependencies/{id}':
    get:
      tags:
        - 'Service instance dependencies'
      summary: 'Display the specified service instance dependency'
      description: 'Get an service instance dependency.'
      operationId: showServiceInstanceDependency
      parameters:
        -
          name: id
          in: path
          description: 'id of the service instance dependency'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Service instance dependency detail'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/resource-service-instance-dependencies' } }
                  message: { type: string }
                type: object
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  exception: { type: string }
                  file: { type: string }
                  line: { type: integer }
                  trace: { type: array, items: { properties: { file: { type: string }, line: { type: integer }, function: { type: string }, class: { type: string }, type: { type: string } }, type: object } }
                type: object
    put:
      tags:
        - 'Service instance dependencies'
      summary: 'Update a service instance dependency'
      description: 'Update a service instance dependency.'
      operationId: updateServiceInstanceDependency
      parameters:
        -
          name: id
          in: path
          description: 'id of the service instance dependency'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request-create-service-instance-dependency'
      responses:
        '200':
          description: 'Updated serviceinstancedependencies data.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/resource-service-instance-dependencies' }
                  message: { type: string }
                type: object
        '422':
          description: 'Invalid inputs.'
          content:
            application/json:
              schema:
                properties:
                  errors: { properties: { '{name-of-invalid-field}': { type: array, items: { description: 'Error message', type: string } } }, type: object }
                  message: { type: string }
                type: object
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  exception: { type: string }
                  file: { type: string }
                  line: { type: integer }
                  trace: { type: array, items: { properties: { file: { type: string }, line: { type: integer }, function: { type: string }, class: { type: string }, type: { type: string } }, type: object } }
                type: object
    delete:
      tags:
        - 'Service instance dependencies'
      summary: 'Delete a service instance dependency'
      description: 'Remove the specified service instance dependency from storage.'
      operationId: deleteServiceInstanceDependency
      parameters:
        -
          name: id
          in: path
          description: 'id of the service instance dependency'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Service instance dependency deleted.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
        '404':
          description: 'Service instance dependency not found.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
...
components:
  schemas:
...
    request-create-service-instance-dependency:
      title: 'Create service instance dependency'
      description: 'Create service instance dependency request'
      required:
        - instance_id
        - instance_dep_id
      properties:
        instance_id:
          description: 'Instance id'
          type: integer
        instance_dep_id:
          description: 'Instance dependency id'
          type: integer
        level:
          description: 'Dependency level (1: minor, 2: major, 3: critical)'
          type: integer
          enum:
            - 1
            - 2
            - 3
        description:
          description: 'Dependency description'
          type: string
      type: object
...
    resource-service-instance-dependencies:
      title: ServiceInstanceDependencies
      description: 'Service instance dependencies resource'
      properties:
        id:
          description: Id
          type: integer
        instance_application_id:
          description: 'Instance application id'
          type: integer
        instance_application_name:
          description: 'Instance application name'
          type: string
        instance_id:
          description: 'Instance id'
          type: integer
        instance_service_name:
          description: 'Instance service name'
          type: string
        instance_dep_application_id:
          description: 'Application id of the instance dependency'
          type: integer
        instance_dep_application_name:
          description: 'Application name of the instance dependency'
          type: string
        instance_dep_id:
          description: 'Instance dependency id'
          type: integer
        instance_dep_service_name:
          description: 'Service name of the instance dependency'
          type: string
        level:
          description: 'Dependency level (1: minor, 2: major, 3: critical)'
          type: integer
          enum:
            - 1
            - 2
            - 3
        description:
          description: 'Dependency description'
          type: string
        created_at:
          description: created_at
          type: string
          format: date-time
        updated_at:
          description: updated_at
          type: string
          format: date-time
      type: object
...
  parameters:
    base-filter-per-page:
      name: perPage
      in: query
      description: 'Number of items displayed by page.'
      schema:
        type: integer
        default: 100
        enum:
          - 5
          - 10
          - 25
          - 100
          - 1000
    base-filter-page:
      name: page
      in: query
      description: 'Current page.'
      schema:
        type: integer
    base-filter-sort:
      name: sort
      in: query
      description: 'Field used to sort the result.'
      schema:
        type: string
      examples:
        string:
          summary: '-id'
          value: '-id'
    base-filter-q:
      name: q
      in: query
      description: 'Query string'
      required: false
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Basic Auth'
      scheme: bearer
security:
  -
    bearerAuth: []